temp_test/backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
celery==5.3.6
channels==4.0.0
channels-redis==4.1.0
dj-database-url==2.1.0
django-cors-headers==4.3.1
django-csp==3.7
PyJWT==2.8.0
python-telegram-bot==20.6
aiohttp==3.9.1
uvicorn==0.27.0
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
whitenoise==6.6.0
Pillow==10.1.0
drf-yasg==1.21.7
cryptography==41.0.4
netaddr==0.8.0 